refactor(src): 🔨 Rename folder artefact to commands
Rename folder artefact to commands.
