+++ src/commands/namepipe.rs
@@ -13,7 +13,7 @@ use std::fs::File;
 use std::path::PathBuf;
 use std::{thread, time};
 
-use crate::artefact::tools::{
+use crate::commands::tools::{
     pretty_print_hashset, regex_to_string, EXIST_ALL_GOOD, EXIST_CLI_ERROR,
 };
 use clap::Parser;
